Fix recommended follow-up commands outputted to the logs
In a number of places in the CLI, we recommend follow-up
`gh [gei|bbs2gh|ado2gh]` commands to run.

When the CLI is run as a `gh` extension rather than
standalone, the examples commands we output in the logs
are incorrect, repeating `gh` twice (e.g. `gh gh-gei`).

This fixes it so we produce accurate, usable commands that
you can actually run.

I chose to fix this in `CliContext.RootCommand` (as
[suggested][1] by @dylan-smith). It would also have been
possible to fix this in the `Program` classes that set
`CliContext.RootCommand`, but that would have required
duplicating logic in 3 places.

Fixes https://github.com/github/gh-gei/issues/1033.