Fix recommended follow-up commands outputted to the logs by statically setting the root command for each CLI
In a number of places in the CLI, we recommend follow-up
`gh [gei|bbs2gh|ado2gh]` commands to run.

When the CLI is run as a `gh` extension rather than standalone, the
examples commands we output in the logs are incorrect, repeating
`gh` twice (e.g.  `gh gh-gei`).

This fixes it so we produce accurate, usable commands that you can
actually run.

We do this by hard-coding the "root command" as a static string
in each `Program` class, rather than trying to deduce it.

Fixes https://github.com/github/gh-gei/issues/1033.