Drop support for non-standard, deprecated `AWS_ACCESS_KEY` and `AWS_SECRET_KEY` environment variables
Up until #855, we used non-standard environment variables for
configuring AWS S3 credentials. With #855, we *added* support for
the standard `AWS_ACCESS_KEY_ID` and `AWS_SECRET_ACCESS_KEY`
variables, but maintained support for the old names, printing a
warning when they're used.

We've had that behaviour for 4 months now, and it's time to cut
the old behaviour.

With this change, only the new environment variable names will
be accepted, and trying to use the old names will result in
your argument not being recognised.

Closes #856.