Require the Bitbucket Server URL, project and repo name to always be passed for `bbs2gh migrate-repo`
This makes a breaking change to `bbs2gh migrate-repo` requiring
the `--bbs-server-url`, `--bbs-project` and `--bbs-repo` to always
be passed, even in the flows where the CLI doesn't talk to the
Bitbucket Server API.

This change is important because it allows us to always sent the
URL of the repository to the GitHub API when starting the
migration, which allows us to intelligently pick the right repo
to migrate from the archive.

Fixes https://github.com/github/gh-gei/issues/1056.