Fix log messages when running `reclaim-mannequin`
The `reclaim-mannequin` command allows reclaiming a single
mannequin like this:

```bash
gh gei reclaim-mannequin --github-org caffeinesoftware \
--mannequin-user redacteduser --target-user timrogers \
--verbose
```

Once the operation has completed, we log something like this:

> [2023-07-19 14:13:23] [INFO] Mannequin reclaim invitation email successfully sent to: redacteduser (M_kgDOCDLDDQ) for timrogers (MDQ6VXNlcjExNjEzNA==)

This log message is the wrong way round - in this case, I
reclaimed a mannequin called `redacteduser` and linked it to the
GitHub user `timrogers`, and the email is going to be sent to me,
not to `redacteduser`.

This fixes that message so it says the right thing. In this PR,
I also take the opportunity to:

* stop logging users' GraphQL IDs which aren't very useful
* try to make the language in the logs clearer to help customers
  to understand exactly what has happened

Fixes https://github.com/github/gh-gei/issues/1072.