BREAKING CHANGE: Drop support for deprecated `--wait` option in `migrate-repo` and `migrate-org`
Up until April, the default behavior of `migrate-repo` and
`migrate-org` was to exit immediately after queueing the
migration, and not to wait for it to complete. Users could opt
in to the "wait" behaviour using the `--wait` option.

In April, we [flipped][1] these defaults around, making waiting
the default and adding the `--queue-only` option for the old
behavior.

This meant that the `--wait` option could go away - but we wanted
to do this in a user-friendly, backwards-compatible way.

At that time, we deprecated the `--wait` option, hiding it from
the in-CLI documentation and printing warning when it's used.
These warnings have been in place for 3 months, so now is a
reasonable time to rip off the band aid and remove the argument.

After this change is released, specifying the old `--wait` option
will cause the CLI to error.

Fixes https://github.com/github/gh-gei/issues/1082 and
https://github.com/github/gh-gei/issues/1059.

[1]: https://github.com/github/gh-gei/pull/929