Fix target repo existence check in GHES migrations so it doesn't explode if the target repo used to exist, but has been renamed
When migrating from GHES with `migrate-repo`, we perform a quick
validation check before starting the export to ensure that the
target repo doesn't already exist. If it does already exist, we
can fail fast and not run the export at all.

If the repo used to exist but has been renamed, then this causes
the CLI to error, because the REST API returns `301 Moved
Permanently` and that's not something we handle or expect.

This updates our code to handle `301 Moved Permanently` and treat
it correctly as "the repo doesn't exist".

Fixes https://github.com/github/gh-gei/issues/1074.