Fail fast if the target GitHub repo already exists in `bbs2gh migrate-repo`
Looking at our error logs, it's clear that customers frequently
start Bitbucket migrations, waiting for the export to run, and
then discover that the  migration can't happen because the target
repo already exists.

This wastes a bunch of time, since Bitbucket archive generation
can be reasonably slow!

This changes the behaviour of `bbs2gh migrate-repo` to match
what `gei migrate-repo` does for GHES migrations: check if the
target repo already exists *before* running the export.

This should only happen when `migrate-repo` is run with a
combination of arguments that includes GitHub credentials - so
not the export-only or the export-and-download modes.

Fixes https://github.com/github/gh-gei/issues/1086.