Extend timeout from 10 hours to 20 hours when generating GitHub Enterprise Server migration archives in `gh gei migrate-repo`
When migrating from GitHub Enterprise Server with `gh gei migrate-
repo`, we give up if the migration archive takes more than 10
hours to generate, and retry.

This may be too aggressive in certain cases, for particularly
large repos or GitHub Enterprise Server instances with limited
resources.

This extends the timeout to 20 hours to allow more time.