Redact `?token=` querystring parameter from GHES <3.8 archive URLs in logs
Existing code allows us to redact secrets in the logs outputted
to the shell and to log files. Secret strings are registered,
and then replaced in the output.

This PR also allows classes using `OctoLogger` to register
regular expressions to replace, and begins redacting
GHES download tokens for GHES <3.8.

Fixes #1151.